fyp-be/notif-scheduler/requirements.txt
psycopg2-binary==2.9.6
rocketry==2.5.1
pydantic==1.10.9
firebase-admin==6.1.0
